+++ tests/test_wordcloud_utils.py
@@ -1,4 +1,4 @@
-from wordcloud_utils import WordCloudCache, generate_wordcloud
+from utilities.wordcloud_utils import WordCloudCache, generate_wordcloud
 
 
 def test_wordcloud_generation_and_retrieval(tmp_path):
